fix(core): external-review hardening — lerp16 UB, scoped crossfade, raw-source viz
Fixes from an on-branch external review (UBSan-assisted):

- render16 lerp16(): the delta × frac intermediate spans ±65535², overflowing
  int32 — signed-overflow UB that happened to produce correct values (and
  passed the extreme-value tests doing it). Intermediate widened to int64;
  descending-direction extremes added to test_render16.
- Mode crossfade is now scoped to the CHANGED segment's window instead of
  blending the whole canvas: other segments render live during the 2 s fade
  rather than easing against a frozen frame. A second effect change mid-fade
  widens the window and re-freezes at the current visual state. New
  discriminating multi-segment test (fails against the whole-canvas blend).
- Viz readback (GET /api/v2/pixels) now mirrors raw sources: sACN and the
  /api/pixels direct path bypass the 16-bit pipeline, so smooth_ held a stale
  ghost of the last native effect. When pipeline16_ is false the endpoint
  serves leds[] verbatim (raw bytes never passed gamma/correction — already
  screen-appropriate). Asserted in the direct-pixels test; API_V2.md updated.
- Command::setEffect removed (enum case, union field, factory): it carried a
  borrowed const char* — a dangling-pointer trap with zero remaining
  producers; effect changes ride ApplyEffectSpec, which owns its id.
- Dead 8-bit-era constants purged: LED_MIN_OUTPUT + LED_CORRECTION_FULL_AT
  survived only in comments describing a floor the 16-bit pipeline explicitly
  removed; comment blocks in constants.h/render16.h/controller.h rewritten to
  the current design.
- show()-rate comments made honest about scaling: ~30 µs/pixel of blocking
  wire time means 'several hundred Hz' holds at ≲300 px but caps near 33 Hz
  at 1000 px — showHz_ reports the truth.
- LED_STRIP_TYPE / LED_COLOR_MODE now #ifndef-guarded so board envs can
  retarget hardware via build_flags (-DLED_STRIP_TYPE=SM16703 ...) instead of
  editing global defaults.

82/82 native tests pass. Not addressed by design: queue-overflow drops the
oldest command (existing policy), trail-effect canvas state during crossfades
(pre-existing), render-RAM cost of the 16-bit buffers (accepted trade).

+++ src/core/render16.h
@@ -20,9 +20,9 @@
 //
 // Everything here is pure integer math with NO FastLED dependency (mirroring
 // core/transition.h), so it is fully host-tested in `pio test -e native`
-// (test_render16) with zero hardware. Hardware-specific final steps — the
-// WS2812B colour correction and the LED_MIN_OUTPUT red-floor — stay in the
-// output driver, downstream of the 8-bit values this produces.
+// (test_render16) with zero hardware. Hardware-specific final shaping (the
+// WS2812B colour correction) lives in the controller's output stage
+// (renderOutput16), downstream of these primitives.
 //
 // Design notes:
 //  • Linear space. Content, compositing, master brightness, and dim-to-warm all
@@ -74,9 +74,14 @@ inline uint16_t scale16(uint16_t v, uint16_t s) {
     return (uint16_t)(((uint32_t)v * (s + 1)) >> 16);
 }
 
-// Linear interpolate two 16-bit channels; frac 0 → a, 65535 → b.
+// Linear interpolate two 16-bit channels; frac 0 → a, 65535 → b. The delta ×
+// frac product spans ±65535², which overflows int32 — signed overflow is UB
+// (UBSan-verified; the old code only *happened* to produce correct values).
+// Widen the intermediate to int64: one 64-bit multiply per channel, and this
+// only runs during mode crossfades.
 inline uint16_t lerp16(uint16_t a, uint16_t b, uint16_t frac) {
-    return (int32_t)a + (((int32_t)b - (int32_t)a) * (int32_t)frac >> 16);
+    return (uint16_t)((int32_t)a +
+                      (int32_t)((((int64_t)b - (int64_t)a) * (int64_t)frac) >> 16));
 }
 
 inline CRGB16 blend16(const CRGB16& a, const CRGB16& b, uint16_t frac) {
